+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Build test for PHP 8.1
 ### Changed
+- The `MarkupFixer` now prefixes `toc-` in cases where HTML tags are encountered with IDs that begin with a decimal
+  (thanks @ArdaGnsrn!)
 - Upgraded PHPStan to v1.4
 
 ## [3.0.2] - 2021-10-26